+++ src/main/java/site/goldenticket/common/security/authentication/SecurityAuthenticationSuccessHandler.java
@@ -8,26 +8,22 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import site.goldenticket.common.redis.service.RedisService;
 import site.goldenticket.common.response.CommonResponse;
 import site.goldenticket.common.security.authentication.dto.AuthenticationToken;
-import site.goldenticket.common.security.authentication.dto.Token;
-import site.goldenticket.common.security.exception.SaveTokenException;
+import site.goldenticket.common.security.authentication.token.TokenService;
 
 import java.io.IOException;
 import java.util.UUID;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static site.goldenticket.common.response.ErrorCode.SAVE_REFRESH_TOKEN_FAILED;
 
 @Slf4j
 @RequiredArgsConstructor
 public class SecurityAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
 
     private final ObjectMapper objectMapper;
-    private final TokenProvider tokenProvider;
-    private final RedisService redisService;
+    private final TokenService tokenService;
 
     @Override
     public void onAuthenticationSuccess(
@@ -39,31 +35,15 @@ public class SecurityAuthenticationSuccessHandler implements AuthenticationSucce
         log.info("Authentication Name = {}", email);
 
         String randomToken = UUID.randomUUID().toString();
-        Token token = tokenProvider.generateToken(randomToken, email);
+        AuthenticationToken authenticationToken = tokenService.generatedToken(randomToken, email);
 
-        try {
-            redisService.set(randomToken, email, token.refreshTokenExpired());
-        } catch (Exception e) {
-            log.error(e.getMessage());
-            throw new SaveTokenException(SAVE_REFRESH_TOKEN_FAILED);
-        }
-
-        sendResponse(response, token);
+        sendResponse(response, authenticationToken);
     }
 
-    private void sendResponse(HttpServletResponse response, Token token) throws IOException {
+    private void sendResponse(HttpServletResponse response, AuthenticationToken authenticationToken) throws IOException {
         response.setStatus(HttpStatus.OK.value());
         response.setContentType(APPLICATION_JSON_VALUE);
         response.setCharacterEncoding(UTF_8.name());
-        objectMapper.writeValue(response.getWriter(), CommonResponse.ok(createAuthenticationToken(token)));
-    }
-
-    private AuthenticationToken createAuthenticationToken(Token token) {
-        return AuthenticationToken.builder()
-                .grantType(token.grantType())
-                .refreshToken(token.refreshToken())
-                .accessToken(token.accessToken())
-                .expiresIn(token.accessTokenExpired())
-                .build();
+        objectMapper.writeValue(response.getWriter(), CommonResponse.ok(authenticationToken));
     }
 }
