+++ src/main/java/site/goldenticket/common/security/authentication/dto/AuthenticationToken.java
@@ -1,6 +1,7 @@
 package site.goldenticket.common.security.authentication.dto;
 
 import lombok.Builder;
+import site.goldenticket.common.security.authentication.token.dto.Token;
 
 @Builder
 public record AuthenticationToken(
@@ -9,4 +10,13 @@ public record AuthenticationToken(
         Long expiresIn,
         String accessToken
 ) {
+
+    public static AuthenticationToken of(Token token) {
+        return AuthenticationToken.builder()
+                .grantType(token.grantType())
+                .refreshToken(token.refreshToken())
+                .accessToken(token.accessToken())
+                .expiresIn(token.accessTokenExpired())
+                .build();
+    }
 }
