+++ src/main/java/site/goldenticket/common/security/authentication/token/TokenProvider.java
@@ -1,4 +1,4 @@
-package site.goldenticket.common.security.authentication;
+package site.goldenticket.common.security.authentication.token;
 
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
@@ -7,7 +7,7 @@ import io.jsonwebtoken.security.SignatureException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import site.goldenticket.common.security.authentication.dto.Token;
+import site.goldenticket.common.security.authentication.token.dto.Token;
 
 import javax.crypto.SecretKey;
 import java.security.Key;
