+++ src/main/java/site/goldenticket/common/security/authentication/token/TokenService.java
@@ -1,0 +1,8 @@
+package site.goldenticket.common.security.authentication.token;
+
+import site.goldenticket.common.security.authentication.dto.AuthenticationToken;
+
+public interface TokenService {
+
+    AuthenticationToken generatedToken(String randomToken, String username);
+}
